fix: lazy-import transformers/accelerate in cosmos_predict2, add CI deps
Move transformers and accelerate imports from module-level to
CosmosPredict2Pipeline.__init__, where they are actually used.
This allows pure key-mapping functions (_remap_state_dict_keys,
_split_state_dict_keys) to be imported without heavy ML deps.

Add torchvision, einops, numpy to CI pip install step — required
by cosmos_predict2_modeling.py.

Also: default attention_backend to 'torch' to avoid
transformer_engine import in test environments.

+++ optimizers/generic_optim.py
@@ -310,7 +310,7 @@ class GenericOptim(Optimizer):
             correct_dim=False,
             cpu_offload=False,
             kahan_buffer_offload=False,
-            kahan_shift_dtype='bf16',
+            kahan_shift_dtype='fp32',
             muon=False,
             adamuon=False,
             normuon=False,
